+++ src/app/(school)/(reviews)/@filter/course/loading.tsx
@@ -1,4 +1,4 @@
-import { ReviewType } from "@/generated/prisma/client";
+import { ReviewType } from "@/generated/prisma/enums";
 
 import { FilterToggleSection } from "@/common/components/FilterToggleSection";
