+++ src/app/(school)/(reviews)/@rating/professor/[slug]/page.tsx
@@ -1,4 +1,4 @@
-import { ReviewType } from "@/generated/prisma/client";
+import { ReviewType } from "@/generated/prisma/enums";
 
 import { auth } from "@/server/auth";
 import { RatingSection } from "@/modules/reviews/components/RatingSection";
