+++ src/app/(school)/bidding/analytics/page.tsx
@@ -9,8 +9,7 @@ import {
 import { Separator } from "@/common/components/separator";
 import { BidPredictionCard } from "@/modules/bidding/components/BidPredictionCard";
 import { notFound } from "next/navigation";
-import { PredictionType } from "@/generated/prisma/client";
-import type { UniversityAbbreviation } from "@/generated/prisma/client";
+import { PredictionType, type UniversityAbbreviation } from "@/generated/prisma/enums";
 import { ModAlternativesCard } from "@/modules/bidding/components/ModAlternativesCard";
 import { BidAnalyticsClient } from "@/modules/bidding/components/BidAnalyticsClient";
 import { AddToTimetableButton } from "@/modules/bidding/components/AddToTimetableButton";
