+++ src/modules/opengraph/components/OgImageHeader/OgImageHeader.tsx
@@ -1,6 +1,6 @@
 import { AfterclassIcon, CustomIcon } from "@/common/components/icons";
 import { NTUIcon, NUSIcon, SMUIcon } from "@/common/components/icons";
-import { type UniversityAbbreviation } from "@/generated/prisma/client";
+import { type UniversityAbbreviation } from "@/generated/prisma/enums";
 
 const svgs = {
   SMU: SMUIcon,
