+++ src/modules/reviews/components/FilterToggleSection/FilterToggleSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { useState } from "react";
 import { usePathname, useRouter, useSearchParams } from "next/navigation";
-import { type ReviewType } from "@/generated/prisma/client";
+import { type ReviewType } from "@/generated/prisma/enums";
 
 import { FilterToggleSection as Filter } from "@/common/components/FilterToggleSection";
 import { type FilterItem } from "@/common/components/FilterToggleSection/FilterToggleSectionItem";
