+++ src/modules/reviews/components/ReviewItem/ReviewReactionsGroup/ReviewReactionsGroup.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useSession } from "next-auth/react";
-import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/client";
+import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/enums";
 
 import { api } from "@/common/tools/trpc/react";
 import { ReviewReactionType } from "@/modules/reviews/types";
