+++ src/modules/roadmaps/components/RoadmapReactionButton.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/client";
+import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/enums";
 
 import { toTitleCase } from "@/common/functions";
 import { Button } from "@/common/components/button";
