+++ src/modules/roadmaps/components/RoadmapReactionsGroup.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/client";
+import { type ReviewReactionType as DbReviewReactionType } from "@/generated/prisma/enums";
 
 import { api } from "@/common/tools/trpc/react";
 import { ReviewReactionType } from "@/modules/reviews/types";
