+++ src/modules/search/components/SearchResult/SearchResultItem/SearchResultItem.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useSession } from "next-auth/react";
-import { type UniversityAbbreviation } from "@/generated/prisma/client";
+import { type UniversityAbbreviation } from "@/generated/prisma/enums";
 
 import { ChevronRightIcon, SchoolIcon } from "@/common/components/icons";
 import {
