+++ src/server/api/courses/getAllByUniAbbrv/index.ts
@@ -1,5 +1,6 @@
 import { z } from "zod";
-import { type Prisma, UniversityAbbreviation } from "@/generated/prisma/client";
+import { type Prisma } from "@/generated/prisma/client";
+import { UniversityAbbreviation } from "@/generated/prisma/enums";
 
 import { publicProcedure } from "@/server/api/trpc";
