+++ src/server/api/reviews/getByCourseCodeProtected/index.ts
@@ -1,4 +1,4 @@
-import { ReviewType } from "@/generated/prisma/client";
+import { ReviewType } from "@/generated/prisma/enums";
 import { z } from "zod";
 
 import { protectedProcedure } from "@/server/api/trpc";
