+++ src/server/api/reviews/getMetadataForCourse/index.ts
@@ -1,4 +1,5 @@
-import { ReviewType, type Prisma } from "@/generated/prisma/client";
+import { type Prisma } from "@/generated/prisma/client";
+import { ReviewType } from "@/generated/prisma/enums";
 import { z } from "zod";
 
 import { publicProcedure } from "@/server/api/trpc";
