+++ src/server/api/roadmaps/copyRoadmap.ts
@@ -1,4 +1,4 @@
-import { Visibility } from "@/generated/prisma/client";
+import { Visibility } from "@/generated/prisma/enums";
 import type { PrismaClient } from "@/generated/prisma/client";
 
 const ENTRY_SELECT = {
