+++ src/server/api/roadmaps/listPublic/index.ts
@@ -1,4 +1,5 @@
-import { Visibility, type Prisma } from "@/generated/prisma/client";
+import { type Prisma } from "@/generated/prisma/client";
+import { Visibility } from "@/generated/prisma/enums";
 
 import { publicProcedure } from "@/server/api/trpc";
